+++ eslint.config.mjs
@@ -14,5 +14,5 @@ export default [
       '@typescript-eslint/no-explicit-any': 'warn',
     },
   },
-  globalIgnores(['**/dist/**']),
+  globalIgnores(['**/dist/**', '**/artifacts/**']),
 ]