+++ examples/hardhat3-plugin/ignition/modules/PrivateEvents.ts
@@ -1,0 +1,9 @@
+import { buildModule } from '@nomicfoundation/hardhat-ignition/modules'
+
+const PrivateEvents = buildModule('PrivateEvents', (m) => {
+  const privateEvents = m.contract('PrivateEvents')
+
+  return { privateEvents }
+})
+
+export default PrivateEvents